Document_upload_assistant/requirements.txt
gradio>=4.0.0
httpx>=0.24.0
tenacity>=8.2.0
PyMuPDF>=1.22.0
python-docx>=0.8.11
pandas>=2.0.0
openpyxl>=3.1.0
baidu-aip>=4.16.0 